# apps/go-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TBD",
        "title": "GO API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/categories": {
            "get": {
                "description": "Returns all categories from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "List all categories",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of products to return (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of products to skip (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListCategoryResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new category to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Create a category",
                "parameters": [
                    {
                        "description": "Category data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/categories/count": {
            "get": {
                "description": "Returns count categories from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Count all categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CountResponse"
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "put": {
                "description": "Update a category to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Update a category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Category data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a category to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Delete a category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteCategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns health check status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Information about health status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HealthResponse"
                        }
                    }
                }
            }
        },
        "/resident-location": {
            "get": {
                "description": "Returns all resident location from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resident_location"
                ],
                "summary": "List all resident location",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of products to return (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of products to skip (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListResidentLocationResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new resident location to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resident_location"
                ],
                "summary": "Create a resident location",
                "parameters": [
                    {
                        "description": "Resident Location data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateResidentLocationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.ResidentLocationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/resident-location/count": {
            "get": {
                "description": "Returns count resident location from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resident_location"
                ],
                "summary": "Count all resident location",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CountResponse"
                        }
                    }
                }
            }
        },
        "/resident-location/{id}": {
            "put": {
                "description": "Update a new resident location to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resident_location"
                ],
                "summary": "Update a resident location",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resident Location ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Resident Location data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateResidentLocationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.ResidentLocationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a new resident location to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resident_location"
                ],
                "summary": "Delete a resident location",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resident Location ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteResidentLocationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CategoryResponse": {
            "type": "object",
            "required": [
                "created_at",
                "description",
                "id",
                "name",
                "updated_at"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-07-09 08:34:05.796617"
                },
                "description": {
                    "type": "string",
                    "example": "Коммунальные платежи"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "utilities"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-07-09 08:34:05.796617"
                }
            }
        },
        "dto.CountResponse": {
            "type": "object",
            "required": [
                "total"
            ],
            "properties": {
                "total": {
                    "type": "integer",
                    "example": 138
                }
            }
        },
        "dto.CreateCategoryRequest": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Коммунальные платежи"
                },
                "name": {
                    "type": "string",
                    "example": "utilities"
                }
            }
        },
        "dto.CreateResidentLocationRequest": {
            "type": "object",
            "required": [
                "apartment",
                "city",
                "country",
                "house",
                "postal_code",
                "street"
            ],
            "properties": {
                "apartment": {
                    "type": "string",
                    "example": "2"
                },
                "city": {
                    "type": "string",
                    "example": "Warsaw"
                },
                "country": {
                    "type": "string",
                    "example": "Poland"
                },
                "house": {
                    "type": "string",
                    "example": "1"
                },
                "postal_code": {
                    "type": "string",
                    "example": "00-945"
                },
                "street": {
                    "type": "string",
                    "example": "Bobr"
                }
            }
        },
        "dto.DeleteCategoryResponse": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.DeleteResidentLocationResponse": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.HealthResponse": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "dto.ListCategoryResponse": {
            "type": "object",
            "required": [
                "data",
                "total"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CategoryResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.ListResidentLocationResponse": {
            "type": "object",
            "required": [
                "data",
                "total"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ResidentLocationResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.ResidentLocationResponse": {
            "type": "object",
            "required": [
                "apartment",
                "city",
                "country",
                "created_at",
                "house",
                "id",
                "postal_code",
                "street",
                "updated_at"
            ],
            "properties": {
                "apartment": {
                    "type": "string",
                    "example": "2"
                },
                "city": {
                    "type": "string",
                    "example": "Warsaw"
                },
                "country": {
                    "type": "string",
                    "example": "Poland"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-07-09 08:34:05.796617"
                },
                "house": {
                    "type": "string",
                    "example": "1"
                },
                "id": {
                    "type": "integer",
                    "example": 123
                },
                "postal_code": {
                    "type": "string",
                    "example": "00-945"
                },
                "street": {
                    "type": "string",
                    "example": "Bobr"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-07-09 08:34:05.796617"
                }
            }
        },
        "dto.UpdateCategoryRequest": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Коммунальные платежи"
                },
                "name": {
                    "type": "string",
                    "example": "utilities"
                }
            }
        },
        "dto.UpdateResidentLocationRequest": {
            "type": "object",
            "required": [
                "apartment",
                "city",
                "country",
                "house",
                "postal_code",
                "street"
            ],
            "properties": {
                "apartment": {
                    "type": "string",
                    "example": "2"
                },
                "city": {
                    "type": "string",
                    "example": "Warsaw"
                },
                "country": {
                    "type": "string",
                    "example": "Poland"
                },
                "house": {
                    "type": "string",
                    "example": "1"
                },
                "postal_code": {
                    "type": "string",
                    "example": "00-945"
                },
                "street": {
                    "type": "string",
                    "example": "Bobr"
                }
            }
        }
    }
}